feat: Added RTC alarm interrupt with deferred processing
