+++ bms/crypto_part/e2studio/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>crypto_bms</name>
+	<name>crypto_bms_IRdriven</name>
 	<comment></comment>
 	<projects>
 	</projects>
